requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-filter==2.0.0
django-haystack==2.8.1
django-js-asset==1.1.0
django-machina==0.7.1
django-mptt==0.9.1
django-rest-swagger==2.2.0
django-tinymce==2.7.0
django-widget-tweaks==1.4.3
djangorestframework==3.8.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
markdown2==2.3.6
MarkupSafe==1.0
oauthlib==2.1.0
openapi-codec==1.3.2
Pillow==5.3.0
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.23
